[iOS] style: 소스 그룹(폴더) 정리
issue: #108
