+++ iOS/AirbnbApp/AirbnbApp/StayDetail/Sections/Models/Amenities.swift
@@ -1,18 +1,5 @@
 import UIKit
 
-class SectionAmenities: ContentView, ViewFromXib {
-    static let xibName = String(describing: SectionAmenities.self)
-
-    @IBOutlet weak var stackView: UIStackView!
-    @IBOutlet weak var viewAllButton: UIButton!
-
-    override func awakeFromNib() {
-        super.awakeFromNib()
-
-        stackView.resetArrangedSubViews()
-    }
-}
-
 struct AmenityCollection {
     private(set) var items: [Amenity]
 
@@ -29,27 +16,6 @@ extension AmenityCollection {
         let amenityTitles = amenitiesText.dropFirst().dropLast().split(separator: ",")
 
         self.items = amenityTitles.map { Amenity(title: String($0), symbol: nil) }
-    }
-}
-
-class SectionAmenitiesFactory {
-    static func makeView(for stayDetail: StayDetail, amenitiesLimit number: Int) -> SectionAmenities {
-        var amenities = AmenityCollection(for: stayDetail)
-        return makeView(for: amenities.limited(number))
-    }
-
-    static func makeView(for amenities: AmenityCollection) -> SectionAmenities {
-        let view = SectionAmenities.loadFromXib()
-
-        amenities.items.map {
-            let view = AmenityView.loadFromXib()
-            view.titleLabel.text = $0.title
-            view.symbolImageView.image = $0.symbol
-
-            return view
-        }.forEach { view.stackView.addArrangedSubview($0) }
-
-        return view
     }
 }
