+++ iOS/AirbnbApp/AirbnbApp/StayDetail/Sections/Models/Summary.swift
@@ -1,0 +1,1 @@
+import Foundation
