+++ iOS/AirbnbApp/AirbnbApp/StayDetail/Sections/Models/TitleInfo.swift
@@ -1,0 +1,15 @@
+import Foundation
+
+struct TitleInfo {
+    let reviewAverage: Float
+    let numberOfReviews: Int
+    let address: String
+}
+
+extension TitleInfo {
+    init(for stayDetail: StayDetail) {
+        self.reviewAverage = stayDetail.reviewInfo.average
+        self.numberOfReviews = stayDetail.reviewInfo.numberOfReviews
+        self.address = stayDetail.locationInfo.address
+    }
+}
