+++ iOS/AirbnbApp/AirbnbApp/StayDetail/Sections/Views/SectionAmenities.swift
@@ -1,0 +1,14 @@
+import UIKit
+
+class SectionAmenities: ContentView, ViewFromXib {
+    static let xibName = String(describing: SectionAmenities.self)
+
+    @IBOutlet weak var stackView: UIStackView!
+    @IBOutlet weak var viewAllButton: UIButton!
+
+    override func awakeFromNib() {
+        super.awakeFromNib()
+
+        stackView.resetArrangedSubViews()
+    }
+}
