+++ iOS/AirbnbApp/AirbnbApp/StayDetail/Sections/Views/SectionAmenities.xib
