+++ iOS/AirbnbApp/AirbnbApp/StayDetail/Sections/Views/SectionBriefInfo.xib
