+++ iOS/AirbnbApp/AirbnbApp/StayDetail/Sections/Views/SectionSummary.xib
