+++ iOS/AirbnbApp/AirbnbApp/StayDetail/Sections/Views/SectionTitleInfo.swift
@@ -1,0 +1,8 @@
+import UIKit
+
+final class SectionTitleInfo: ContentView, ViewFromXib {
+    static let xibName = String(describing: SectionTitleInfo.self)
+
+    @IBOutlet weak var reviewsLabel: UILabel!
+    @IBOutlet weak var addressLabel: UILabel!
+}
